Backend/requirements.txt
APScheduler

Django
django-allauth==64.1.0
django-background-tasks==1.2.8
django-cors-headers==4.4.0
django-environ==0.12.0
django-filter==24.2
django-redis==5.4.0
django-shortcuts==1.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0

environ==1.0
httplib2==0.20.2
hyperlink==21.0.0

psycopg2
psycopg2
psycopg2-binary

PyJWT==2.10.1
pyOpenSSL==23.0.0
pyparsing==3.2.1
redis==6.2.0
requests

